% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promotion_impact.R
\name{create.dummy.vars}
\alias{create.dummy.vars}
\title{generate data.frame with dummy variables for each promotion period}
\usage{
create.dummy.vars(target.data, promotion.data, tovar.col = "pro_id")
}
\description{
generate data.frame with dummy variables for each promotion period
}
\keyword{internal}
