% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promotion_impact.R
\docType{data}
\name{sim.data}
\alias{sim.data}
\title{Daily Total Sales}
\format{A dataset containing 958 observations of 2 variables.}
\source{
NCsoft AnalysisModeling Team <gimmesilver@ncsoft.com> <windy0126@ncsoft.com> <nhkim1302@ncsoft.com>
}
\usage{
sim.data
}
\description{
This data set is simulated daily total sales data contaning 958 observations of 2 variables.
`dt`: date with Date format.
`simulated_sales`: simulated daily sales with numeric format.
}
\keyword{datasets}
