% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promotion_impact.R
\docType{data}
\name{sim.promotion}
\alias{sim.promotion}
\title{Promotion Schedule}
\format{A dataset containing 50 observations of 4 variables.}
\source{
NCsoft AnalysisModeling Team <gimmesilver@ncsoft.com> <windy0126@ncsoft.com> <nhkim1302@ncsoft.com>
}
\usage{
sim.promotion
}
\description{
This data set is promotion schedule data including promotion tag information.
`pro_id`: promotion ID.
`start_dt`: start date of each promotion
`end_dt`: end date of each promotion.
`tag_info`: promotion tag information (promotion type).
}
\keyword{datasets}
