% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promotion_impact.R
\docType{data}
\name{sim.promotion.sales}
\alias{sim.promotion.sales}
\title{Daily Promotion Sales with Promotion information}
\format{A dataset containing 1486 observations of 6 variables.}
\source{
NCsoft AnalysisModeling Team <gimmesilver@ncsoft.com> <windy0126@ncsoft.com> <nhkim1302@ncsoft.com>
}
\usage{
sim.promotion.sales
}
\description{
This data set is simulated daily promotion sales data with promotion information.
`pro_id`: promotion ID
`start_dt`: start date of each promotion
`end_dt`: end date of each promotion
`tag_info`: promotion tag information (promotion type)
`dt`: date
`payment`: simulated daily promotion sales
}
\keyword{datasets}
