% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promotion_impact.R
\name{value_fit_plot}
\alias{value_fit_plot}
\title{generate ggplot with fitted and target values}
\usage{
value_fit_plot(data)
}
\description{
generate ggplot with fitted and target values
}
\keyword{internal}
