% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promotion_impact.R
\name{detectOutliers}
\alias{detectOutliers}
\title{detect some outliers}
\usage{
detectOutliers(model, threshold = list(cooks.distance = 1, dfbetas = 1,
  dffits = 2), option = 2)
}
\arguments{
\item{model}{Execution result object : promotionImpact}

\item{threshold}{List of threshold values to be determined as outliers if greater than the written values}

\item{option}{The number of indicators that must be greater than the threshold values to be outliers.}
}
\description{
detectOutliers
}
\details{
detectOutliers extracts outliers which affect the average effects of promotions.
}
\examples{
pri1 <- promotionImpact(data = sim.data, promotion = sim.promotion.sales, 
                                  time.field = 'dt', target.field = 'simulated_sales')
out <- detectOutliers(model = pri1, 
                      threshold = list(cooks.distance=1, dfbetas=1, dffits=2), option = 1)
out$outliers
sim.data.new <- sim.data[sim.data$dt != '2017-04-02', ]
sim.promotion.sales.new <- sim.promotion.sales[sim.promotion.sales$dt != '2017-04-02', ]
pri2 <- promotionImpact(data = sim.data.new, promotion = sim.promotion.sales.new, 
                        time.field = 'dt', target.field = 'simulated_sales')
pri1$effects
pri2$effects
}
